refactor side events
