+++ src/side-events/clipboard.ts
@@ -1,0 +1,20 @@
+import { clipboard, ClipboardPermissionChangeAction } from '../actions'
+
+export const addPermissionChangeListeners = (
+  dispatch: (action: ClipboardPermissionChangeAction) => void
+) => {
+  clipboard.onReadPermissionChange = permissionState =>
+    dispatch({
+      type: 'clipboardPermissionChange',
+      payload: {
+        read: permissionState
+      }
+    })
+  clipboard.onWritePermissionChange = permissionState =>
+    dispatch({
+      type: 'clipboardPermissionChange',
+      payload: {
+        write: permissionState
+      }
+    })
+}
