+++ src/side-events/index.ts
@@ -1,0 +1,6 @@
+import { addPermissionChangeListeners } from './clipboard'
+import { Action } from '../actions'
+
+export const applySideEventListeners = (dispatch: (action: Action) => void) => {
+  addPermissionChangeListeners(dispatch)
+}
